Make CliIoSetup public and rename it.
